+++ .github/workflows/clang-format.yaml
@@ -24,7 +24,7 @@ jobs:
         run: |
           find include -type f \( -name '*.hpp' -o -name '*.h' \) -exec clang-format -i --style=file --verbose {} +
           find src -type f \( -name '*.hpp' -o -name '*.h' -o -name '*.cpp' \) -exec clang-format -i --style=file --verbose {} +
-          find test -type f \( -name '*.hpp' -o -name '*.h' -o -name '*.cpp' \) ! -path 'test/*' -exec clang-format -i --style=file --verbose {} +
+          find test -type f \( -name '*.hpp' -o -name '*.h' -o -name '*.cpp' \) -exec clang-format -i --style=file --verbose {} +
 
       - name: Check for changes
         id: check-changes
