+++ include/miam/processes.hpp
@@ -3,8 +3,7 @@
 
 #pragma once
 
-#include <miam/aerosol_property.hpp>
-#include <miam/process_set.hpp>
+#include <miam/processes/process_set.hpp>
 #include <miam/processes/constants/equilibrium_constant.hpp>
 #include <miam/processes/constants/henrys_law_constant.hpp>
 #include <miam/processes/dissolved_reaction.hpp>
