+++ include/miam/representations/aerosol_property.hpp
