+++ src/version.hpp.in
@@ -1,5 +1,6 @@
-// Copyright (C) 2026 National Center for Atmospheric Research
+// Copyright (C) 2026 University Corporation for Atmospheric Research
 // SPDX-License-Identifier: Apache-2.0
+//
 // clang-format off
 #pragma once
