+++ test/unit/aerosol_property.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2026 University Corporation for Atmospheric Research
 // SPDX-License-Identifier: Apache-2.0
 
-#include <miam/aerosol_property.hpp>
+#include <miam/representations/aerosol_property.hpp>
 #include <micm/util/vector_matrix.hpp>
 
 #include <map>
