+++ test/unit/model.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2026 University Corporation for Atmospheric Research
 // SPDX-License-Identifier: Apache-2.0
 
-#include <miam/model.hpp>
+#include <miam/model/model.hpp>
 #include <miam/representations/single_moment_mode.hpp>
 #include <miam/representations/two_moment_mode.hpp>
 #include <miam/representations/uniform_section.hpp>
